Fix saving 4GB+ files file size cutting off issue
This commit fixes invalid 64-bit to 32-bit castings of filesize integers.

Fix #12526, close #12736

+++ PowerEditor/src/Utf8_16.cpp
@@ -299,7 +299,7 @@ bool Utf8_16_Write::openFile(const TCHAR *name)
 	return true;
 }
 
-bool Utf8_16_Write::writeFile(const void* p, unsigned long _size)
+bool Utf8_16_Write::writeFile(const void* p, size_t _size)
 {
     // no file open
 	if (!m_pFile)
